+++ gulpfile.js
@@ -1,33 +1,33 @@
 (function () {
     "use strict";
 
-    var gulp = require('gulp'),
-        fs = require('fs'),
+    var fs = require('fs'),
         path = require('path'),
+        gulp = require('gulp'),
         extend = require('extend'),
+        requireAll = require('require-all'),
 
-        configLocation = path.resolve(__dirname, 'tasks/local.json'),
+        configLocation = path.resolve(__dirname, 'tasks/config.json'),
         config = extend({
             port: 3000,
             livereload: true,
             appRoot: '/usr/docs/rise-js/',
-            bumpVersionType: 'prerelease',
             sourceMap: [
-                "src/Rise.js",
-                "src/util/Util.js",
-                "src/util/**/*.js",
-                "src/ElementFactory.js",
-                "src/element/BasicElement.js",
-                "src/element/**/*.js",
-                "src/extension/Extension.js",
-                "src/extension/**/*.js",
-                "src/**/*.js"
+                //"src/Rise.ts",
+                //"src/util/Util.ts",
+                //"src/util/**/*.ts",
+                //"src/ElementFactory.ts",
+                "src/element/BasicElement.ts",
+                "src/element/**/*.ts"
+                //"src/extension/Extension.ts",
+                //"src/extension/**/*.ts",
+                //"src/**/*.ts"
             ].map(function (item) {
                     return path.resolve(__dirname, item);
                 })
         }, fs.existsSync(configLocation) ? require(configLocation) : {}),
 
-        tasks = require('require-all')({
+        tasks = requireAll({
             dirname: path.resolve(__dirname, 'tasks'),
             filter: /(.+)\.js$/,
             excludeDirs: /^\.(git|svn)$/