+++ src/element/EmbedElement.js
@@ -1,5 +1,0 @@
-(function () {
-    "use strict";
-
-    Rise.EmbedElement = Rise.BasicElement.extend();
-}());