+++ src/element/EmbedElement.ts
@@ -1,0 +1,7 @@
+class EmbedElement extends BasicElement {
+    name:string = 'Embed';
+
+    constructor() {
+        super();
+    }
+}