+++ src/element/ImageElement.ts
@@ -1,0 +1,7 @@
+class ImageElement extends BasicElement {
+    name:string = 'Image';
+
+    constructor() {
+        super();
+    }
+}