+++ src/element/TextElement.ts
@@ -1,0 +1,7 @@
+class TextElement extends BasicElement {
+    name:string = 'Text';
+
+    constructor() {
+        super();
+    }
+}