+++ tasks/build-js.js
@@ -4,19 +4,18 @@
     var concat = require('gulp-concat'),
         uglify = require('gulp-uglify'),
         sourcemaps = require('gulp-sourcemaps'),
-        rename = require('gulp-rename');
+        tsc = require('gulp-tsc');
 
     module.exports = function (gulp, config) {
         gulp.task('build-js', ['clean'], function () {
             return gulp.src(config.sourceMap)
-                .pipe(concat('rise.concat.js'))
-                .pipe(gulp.dest('./dist'))
-                .pipe(rename('rise.min.js'))
                 .pipe(sourcemaps.init())
-                .pipe(uglify({
-                    mangle: false
+                .pipe(tsc({
+                    target: "ES5"
                 }))
-                .pipe(sourcemaps.write('./'))
+                .pipe(concat('rise.min.js'))
+                .pipe(uglify())
+                .pipe(sourcemaps.write())
                 .pipe(gulp.dest('./dist'));
         });
     };