+++ tasks/build.js
@@ -1,7 +1,0 @@
-(function () {
-    "use strict";
-
-    module.exports = function (gulp) {
-        gulp.task('build', ['clean', 'run-validation', 'run-tests', 'bump-version', 'build-js']);
-    };
-}());