+++ tasks/webserver.js
@@ -5,7 +5,7 @@
         path = require('path');
 
     module.exports = function (gulp, config) {
-        gulp.task('webserver', ['linker-dev'], function () {
+        gulp.task('webserver', function () {
             return gulp.src(path.resolve(__dirname, '../'))
                 .pipe(webserver({
                     port: config.port,